refactor: extract `header`, `padding`, `border`, `margin` from `theme`
* remove `seeMoreTitle`
* move `padding`, `margin`, `border` to css
