+++ apps/calendar/src/css/popup/common.css
@@ -5,7 +5,7 @@
   background: none;
   border: 0;
   font-size: 14px;
-  line-height: 3.6;
-  padding: 0 17px;
+  line-height: 28px;
+  padding-right: 17px;
   cursor: pointer;
 }
